Create Search
Here the user enters names as arguments before running the program. The program confirms how many names have been stored and then prompts the user to search for a name.
The program iterates over the user entered array to find a match and will tell the user whether the name is found or not found.