Create Favourite
Simple program that allows names to be entered as arguments, that can be voted for by a number of voters (specified by the user) each casting one vote. The program will keep prompting a user until they enter a valid vote that matches one of the names entered initially as an argument. After all votes are cast, the program calculates who has the highest number of votes and announces that name as the favourite